requirements.txt
nltk==3.1
PyYAML==3.11
textblob==0.11.0
wheel==0.24.0
